vpnc: skip parsing responder lifetime payload
My company upgraded its Fortigate firewall recently, and I started seeing this assertion pop when connecting to it. Looking at the debug output from vpnc, I see this:

   PARSING PAYLOAD type: 0b (ISAKMP_PAYLOAD_N)
   next_type: 0b (ISAKMP_PAYLOAD_N)
   length: 0028
   n.doi: 00000001 (ISAKMP_DOI_IPSEC)
   n.protocol: 01 (ISAKMP_IPSEC_PROTO_ISAKMP)
   n.spi_length: 10
   n.type: 6000 (ISAKMP_N_IPSEC_RESPONDER_LIFETIME)
   n.spi: 66548f57 ce7a153b ab19c18b 851d4fbb
   n.data: 800b0001 00020004 00007080
   t.attributes.type: 000b (IKE_ATTRIB_LIFE_TYPE)
   t.attributes.u.attr_16: 0001 (IKE_LIFE_TYPE_SECONDS)
   t.attributes.type: 0002 (IKE_ATTRIB_HASH)
   t.attributes.u.lots.length: 0004
   t.attributes.u.lots.data: 00007080
   DONE PARSING PAYLOAD type: 0b (ISAKMP_PAYLOAD_N)

...which makes little sense to me. The firewall is sending a
IKE_ATTRIB_LIFE_TYPE with a value of seconds in it, but then it sends an
attribute specifying the hash algorithm. I think this is a bug in the
firewall code, and I've asked our IT people to report it to fortinet.

This patch works around the problem. In the event that we get a bogus
responder lifetime payload that doesn't include a value for the
lifetime, just skip parsing it instead of asserting.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
(From: https://bugzilla.redhat.com/show_bug.cgi?id=1171852)